// sbah1-main/types.ts
export type Language = 'ar' | 'fr' | 'en';
export type TranslatableString = { [key in Language]?: string; };

export interface LanguageContextType {
  language: Language;
  setLanguage: (lang: Language) => void;
  t: (key: string, replacements?: Record<string, string | number>) => string;
  dbTranslations: Record<string, Record<string, string>>;
  setDbTranslations: React.Dispatch<React.SetStateAction<Record<string, Record<string, string>>>>;
  isLoadingTranslations: boolean;
  translateField: (field: TranslatableString | string | undefined) => string;
}